# openapi.yaml
openapi: 3.0.3
info:
  title: Finance AI API
  description: API for managing financial transactions and user reports.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid input
        '409':
          description: User already exists

  /api/login:
    post:
      summary: Authenticate user and receive token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials

  /api/transactions:
    get:
      summary: Get all transactions for the authenticated user
      tags:
        - Transactions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '401':
          description: Unauthorized

  /api/upload:
    post:
      summary: Upload a financial report (PDF, XLSX, CSV)
      tags:
        - Transactions
      security:
        - BearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded and processing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized

  /api/uploads:
    get:
      summary: List all file upload batches
      tags:
        - Transactions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of upload batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Upload'
        '401':
          description: Unauthorized

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    Transaction:
      type: object
      properties:
        transaction_id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        source:
          type: string
          example: "brubank"
        account:
          type: string
          example: "caja_ahorro_pesos"
        date:
          type: string
          format: date
          example: "2025-02-06"
        amount:
          type: float
          example: -97380.39
        currency:
          type: string
          example: "ARS"
        description:
          type: string
          example: "Debito automatico SANCOR"
        direction:
          type: string
          enum: [debit, credit]
        merchant:
          type: string
          nullable: true
          example: "SANCOR"
        category:
          type: string
          nullable: true
          example: "seguros"
        subcategory:
          type: string
          nullable: true
          example: "debito automatico"
        balance:
          type: number
          nullable: true
          example: 29613.26
        is_transfer:
          type: boolean
        is_fee:
          type: boolean
        is_tax:
          type: boolean
        upload_id:
          type: string
          format: uuid
          example: "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"

    Upload:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        status:
          type: string
          enum: [processing, completed, failed]
        created_at:
          type: string
          format: date-time